+++ Source/BuildPackages/Templates/Parser_contains.pas
@@ -1,4 +1,5 @@
 contains
   NovusTemplate in '..\..\Core\Parser\NovusTemplate.pas',
   NovusParser in '..\..\Core\Parser\NovusParser.pas',
+  NovusTokenProcessor in '..\..\Core\Parser\NovusTokenProcessor.pas',
   NovusStringParser in '..\..\Core\Parser\NovusStringParser.pas';